backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]
pillow
torch==2.7.0+cu118     --extra-index-url https://download.pytorch.org/whl/cu118
torchvision==0.22.0+cu118  --extra-index-url https://download.pytorch.org/whl/cu118
